seamless-mode: Add support for single window changes
Agent can now send just the coordinates of the window that has changed.
In the case window has been hidden or closed, width of the window must be
set to 0.

This reduces CPU load because the agent doesn't have to cycle through all the
windows and decide which to send. It mainly concerns Windows agent as there
are often hundreds of windows and majority of them is not visible to the user.